+++ stac_fastapi/elasticsearch/tests/conftest.py
@@ -65,6 +65,7 @@ Settings.set(settings)
 @pytest.fixture(scope="session")
 def event_loop():
     loop = asyncio.new_event_loop()
+    asyncio.set_event_loop(loop)
     yield loop
     loop.close()